# backend/tests/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AgentCreate": {
        "additionalProperties": false,
        "properties": {
          "auto_delegate": {
            "default": false,
            "title": "Auto Delegate",
            "type": "boolean"
          },
          "changelog": {
            "default": "",
            "maxLength": 2000,
            "title": "Changelog",
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/VersionConfig"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "handle": {
            "pattern": "^[a-z][a-z0-9_-]{2,31}$",
            "title": "Handle",
            "type": "string"
          },
          "is_callable": {
            "default": true,
            "title": "Is Callable",
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "source_agent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Agent Id"
          },
          "use_cases": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Use Cases",
            "type": "array"
          },
          "version_policy": {
            "default": "pin",
            "enum": [
              "pin",
              "follow_active_at_cycle"
            ],
            "title": "Version Policy",
            "type": "string"
          }
        },
        "required": [
          "handle",
          "name",
          "config"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
      "AgentDetail": {
        "properties": {
          "active_mapping_count": {
            "title": "Active Mapping Count",
            "type": "integer"
          },
          "auto_delegate": {
            "title": "Auto Delegate",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "handle": {
            "title": "Handle",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_callable": {
            "title": "Is Callable",
            "type": "boolean"
          },
          "is_template": {
            "title": "Is Template",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_agent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Agent Id"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "use_cases": {
            "items": {
              "type": "string"
            },
            "title": "Use Cases",
            "type": "array"
          },
          "version_policy": {
            "enum": [
              "pin",
              "follow_active_at_cycle"
            ],
            "title": "Version Policy",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "handle",
          "name",
          "description",
          "use_cases",
          "is_template",
          "source_agent_id",
          "is_callable",
          "auto_delegate",
          "version_policy",
          "created_at",
          "updated_at",
          "status",
          "active_mapping_count",
          "versions"
        ],
        "title": "AgentDetail",
        "type": "object"
      },
      "AgentDuplicate": {
        "additionalProperties": false,
        "properties": {
          "handle": {
            "pattern": "^[a-z][a-z0-9_-]{2,31}$",
            "title": "Handle",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "handle",
          "name"
        ],
        "title": "AgentDuplicate",
        "type": "object"
      },
      "AgentListItem": {
        "properties": {
          "active_mapping_count": {
            "title": "Active Mapping Count",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "handle": {
            "title": "Handle",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_template": {
            "title": "Is Template",
            "type": "boolean"
          },
          "latest_active_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Active Version"
          },
          "latest_version": {
            "title": "Latest Version",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "handle",
          "name",
          "description",
          "is_template",
          "status",
          "latest_version",
          "latest_active_version",
          "active_mapping_count",
          "updated_at"
        ],
        "title": "AgentListItem",
        "type": "object"
      },
      "AgentPatch": {
        "additionalProperties": false,
        "properties": {
          "auto_delegate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Delegate"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_callable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Callable"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "use_cases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Cases"
          },
          "version_policy": {
            "anyOf": [
              {
                "enum": [
                  "pin",
                  "follow_active_at_cycle"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Policy"
          }
        },
        "title": "AgentPatch",
        "type": "object"
      },
...
      "ArchiveResult": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentDetail"
          },
          "deactivated_mapping_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Deactivated Mapping Ids",
            "type": "array"
          }
        },
        "required": [
          "agent",
          "deactivated_mapping_ids"
        ],
        "title": "ArchiveResult",
        "type": "object"
      },
...
      "Capability": {
        "additionalProperties": false,
        "properties": {
          "connector": {
            "title": "Connector",
            "type": "string"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "connector",
          "tools"
        ],
        "title": "Capability",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Page_AgentListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "Page[AgentListItem]",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Recurrence": {
        "additionalProperties": false,
        "properties": {
          "every_days": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Every Days",
            "type": "integer"
          }
        },
        "required": [
          "every_days"
        ],
        "title": "Recurrence",
        "type": "object"
      },
...
      "ValidateRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/VersionConfig"
          }
        },
        "required": [
          "config"
        ],
        "title": "ValidateRequest",
        "type": "object"
      },
      "ValidateResult": {
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "title": "ValidateResult",
        "type": "object"
      },
      "VersionConfig": {
        "additionalProperties": false,
        "properties": {
          "alert_policy": {
            "$ref": "#/components/schemas/AlertPolicy"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "minItems": 1,
            "title": "Capabilities",
            "type": "array"
          },
          "end_conditions": {
            "$ref": "#/components/schemas/EndConditions",
            "default": {
              "max_duration_days": 120,
              "max_steps": 600,
              "on_subject_closed": "end"
            }
          },
          "follow_up": {
            "$ref": "#/components/schemas/FollowUp",
            "default": {
              "ladder": [],
              "mode": "none"
            }
          },
          "hitl": {
            "$ref": "#/components/schemas/Hitl",
            "default": {
              "ask_on": [],
              "verify_evidence_below": 0.8
            }
          },
          "max_turns_per_episode": {
            "default": 12,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Turns Per Episode",
            "type": "integer"
          },
          "models": {
            "$ref": "#/components/schemas/Models"
          },
          "policy_pack": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PolicyRuleRef"
            },
            "title": "Policy Pack",
            "type": "array"
          },
          "recurrence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recurrence"
              },
              {
                "type": "null"
              }
            ]
          },
          "system_prompt": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "System Prompt",
            "type": "string"
          }
        },
        "required": [
          "system_prompt",
          "models",
          "capabilities",
          "alert_policy"
        ],
        "title": "VersionConfig",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lucia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agents": {
      "get": {
        "operationId": "listAgents",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "archived"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "is_template",
            "required": false,
            "schema": {
              "default": false,
              "title": "Is Template",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-updated_at",
              "pattern": "^(name|-name|handle|updated_at|-updated_at|created_at|-created_at)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AgentListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "List agents",
        "tags": [
          "internal:agents"
        ]
      },
      "post": {
        "operationId": "createAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Create an agent and its v1",
        "tags": [
          "internal:agents"
        ]
      }
    },
    "/api/v1/agents/validate": {
      "post": {
        "operationId": "validateAgentConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Validate a version config against the registry",
        "tags": [
          "internal:agents"
        ]
      }
    },
    "/api/v1/agents/{handle}": {
      "get": {
        "operationId": "getAgent",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Get an agent with its versions",
        "tags": [
          "internal:agents"
        ]
      },
      "patch": {
        "operationId": "updateAgent",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Update agent metadata (no new version)",
        "tags": [
          "internal:agents"
        ]
      }
    },
    "/api/v1/agents/{handle}/archive": {
      "post": {
        "operationId": "archiveAgent",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Archive every version and deactivate its mappings",
        "tags": [
          "internal:agents"
        ]
      }
    },
    "/api/v1/agents/{handle}/duplicate": {
      "post": {
        "operationId": "duplicateAgent",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDuplicate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Copy an agent's latest active version into a new agent",
        "tags": [
          "internal:agents"
        ]
      }
    },
    "/api/v1/agents/{handle}/unarchive": {
      "post": {
        "operationId": "unarchiveAgent",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Reactivate the latest version",
        "tags": [
          "internal:agents"
        ]
      }
    },
...
  },
...
}